chore: rename entry-ssr
